Fix missing warnings in CMakeLists and adapt integrators for warnings

+++ transformations/integrator/IntegratorRect.cc
@@ -68,7 +68,7 @@ void IntegratorRect::sample(FunctionArgs& fargs){
   }
 
   size_t offset=0;
-  for (size_t i = 0; i < m_orders.size(); ++i) {
+  for (size_t i = 0; i < static_cast<size_t>(m_orders.size()); ++i) {
     auto n=m_orders[i];
     if(n>1){
       abscissa.segment(offset, n)=ArrayXd::LinSpaced(n, low[i], high[i]);
